functions/process_voice_transcript/requirements.txt
functions-framework==3.*
google-cloud-storage==2.10.*
google-cloud-bigquery==3.11.*
deepgram-sdk==3.*